requirements.txt
langchain==0.2.16
langchain_community==0.2.16
langchain_groq==0.1.9
langchain_openai==0.1.23
langchain_experimental==0.0.65
pypdf==4.3.1
python-dotenv==1.0.1
openai==1.44.0
tiktoken==0.7.0
pandas==2.2.2
openpyxl==3.1.5
langchain_chroma==0.1.4
streamlit==1.38.0
